requirements.txt
click
PyVCF
pysam
